Make engine easier to understand.
This is a first go at making the engine logic easier to follow.
Lets break it down to multiple files and add some docs to what's
happening. Treat this as a first go changing the structure by first
making it a little easier to follow.
