+++ lib/core/seleniumRunner.js
@@ -256,6 +256,71 @@ class SeleniumRunner {
     }
     return Promise.resolve();
   }
+
+  runScriptFromCategory(script, isAsync, name) {
+    // Scripts should be valid statements or IIFEs '(function() {...})()' that can run
+    // on their own in the browser console. Prepend with 'return' to return result of statement to Browsertime.
+    return Promise.resolve(script).then(script => {
+      if (isAsync) {
+        const source = `
+            var callback = arguments[arguments.length - 1];
+            return (${script})
+              .then((r) => callback({'result': r}))
+              .catch((e) => callback({'error': e}));
+            `;
+
+        return this.runAsyncScript(source, name).then(result => {
+          if (result.error) {
+            throw result.error;
+          } else {
+            return result.result;
+          }
+        });
+      } else {
+        const source = 'return ' + script;
+        return this.runScript(source, name);
+      }
+    });
+  }
+
+  runScripts(scriptsByCategory, isAsync) {
+    if (!scriptsByCategory) {
+      return Promise.resolve({});
+    }
+    return Promise.resolve(scriptsByCategory).then(scripts => {
+      const categoryNames = Object.keys(scripts);
+      return Promise.reduce(
+        categoryNames,
+        (results, categoryName) => {
+          const category = scripts[categoryName];
+
+          return this.runScriptInCategory(category, isAsync).then(result => {
+            results[categoryName] = result;
+            return results;
+          });
+        },
+        {}
+      );
+    });
+  }
+
+  runScriptInCategory(category, isAsync) {
+    return Promise.reduce(
+      Object.keys(category),
+      (results, scriptName) => {
+        const script = category[scriptName];
+        return this.runScriptFromCategory(script, isAsync, scriptName).then(
+          result => {
+            if (!(result === null || result === undefined)) {
+              results[scriptName] = result;
+            }
+            return results;
+          }
+        );
+      },
+      {}
+    );
+  }
 }
 
 module.exports = SeleniumRunner;
